Move HistoryConfig to history package to avoid circular references
