+++ config/storage_local_test.go
@@ -7,42 +7,11 @@ import (
 	"github.com/minamijoyo/tfmigrate/history"
 )
 
-func TestLocalStorageConfigNewStorage(t *testing.T) {
-	cases := []struct {
-		desc   string
-		config *LocalStorageConfig
-		ok     bool
-	}{
-		{
-			desc: "valid",
-			config: &LocalStorageConfig{
-				Path: "tmp/history.json",
-			},
-			ok: true,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.desc, func(t *testing.T) {
-			got, err := tc.config.NewStorage()
-			if tc.ok && err != nil {
-				t.Fatalf("unexpected err: %s", err)
-			}
-			if !tc.ok && err == nil {
-				t.Fatalf("expected to return an error, but no error, got: %#v", got)
-			}
-			if tc.ok {
-				_ = got.(*history.LocalStorage)
-			}
-		})
-	}
-}
-
 func TestParseLocalStorageBlock(t *testing.T) {
 	cases := []struct {
 		desc   string
 		source string
-		want   StorageConfig
+		want   history.StorageConfig
 		ok     bool
 	}{
 		{
@@ -57,7 +26,7 @@ tfmigrate {
   }
 }
 `,
-			want: &LocalStorageConfig{
+			want: &history.LocalStorageConfig{
 				Path: "tmp/history.json",
 			},
 			ok: true,
