+++ config/storage_s3_test.go
@@ -7,43 +7,11 @@ import (
 	"github.com/minamijoyo/tfmigrate/history"
 )
 
-func TestS3StorageConfigNewStorage(t *testing.T) {
-	cases := []struct {
-		desc   string
-		config *S3StorageConfig
-		ok     bool
-	}{
-		{
-			desc: "valid",
-			config: &S3StorageConfig{
-				Bucket: "tfmigrate-test",
-				Key:    "tfmigrate/history.json",
-			},
-			ok: true,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.desc, func(t *testing.T) {
-			got, err := tc.config.NewStorage()
-			if tc.ok && err != nil {
-				t.Fatalf("unexpected err: %s", err)
-			}
-			if !tc.ok && err == nil {
-				t.Fatalf("expected to return an error, but no error, got: %#v", got)
-			}
-			if tc.ok {
-				_ = got.(*history.S3Storage)
-			}
-		})
-	}
-}
-
 func TestParseS3StorageBlock(t *testing.T) {
 	cases := []struct {
 		desc   string
 		source string
-		want   StorageConfig
+		want   history.StorageConfig
 		ok     bool
 	}{
 		{
@@ -59,7 +27,7 @@ tfmigrate {
   }
 }
 `,
-			want: &S3StorageConfig{
+			want: &history.S3StorageConfig{
 				Bucket: "tfmigrate-test",
 				Key:    "tfmigrate/history.json",
 			},
