+++ history/storage.go
@@ -15,3 +15,9 @@ type Storage interface {
 	// an empty array instead of an error.
 	Read(ctx context.Context) ([]byte, error)
 }
+
+// StorageConfig is an interface of factory method for Storage
+type StorageConfig interface {
+	// NewStorage returns a new instance of Storage.
+	NewStorage() (Storage, error)
+}
